GREEDO-BANK-T22-fix issues

+++ src/main/java/com/greedobank/reports/dao/NewsDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewsDAO extends JpaRepository<News, Integer> {
+public interface NewsDAO extends JpaRepository<News, Long> {
 }
